requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-ollama
langchain-community
langchain-core
psycopg2-binary
pgvector
sqlalchemy
redis
python-dotenv
pydantic
pypdf
python-multipart